📝 Store a list of all installed plugin:versions
This list obtained from the current image, by going to this path on a
running instance:

http://<jenkins>/pluginManager/api/xml?depth=1&xpath=/*/*/shortName|/*/*/version&wrapper=plugins

From here: https://stackoverflow.com/a/12730830/2829381

I stripped the xml, and formatted the output the same way that the s2i
build lists "Installed plugins" at the end. Future updates can just
take that list as-is. This will then give us a diff of what
plugin:version changes there are between 2 images.