✨ Install lockable-resources plugin
https://plugins.jenkins.io/lockable-resources

This plugin will allow us to prevent jobs from stepping on each other,
if there are shared resources that they need to be the sole consumer
of. Also, from the global config, we can mark the resource as locked,
so that no jobs will attempt to use it -- this will be useful if we
need to do something manual with the resource, and don't want jobs
interfering.